+++ packages/connect-wallet-modal/src/connectButtons/index.ts
@@ -13,6 +13,7 @@ export { default as ConnectZenGo } from './connectZenGo';
 export { default as ConnectBraveWallet } from './connectBraveWallet';
 export { default as ConnectOperaWallet } from './connectOperaWallet';
 export { default as ConnectExodus } from './connectExodus';
+export { default as connectFrontier } from './connectFrontier';
 export { default as ConnectGamestop } from './connectGamestop';
 export { default as ConnectXdefi } from './connectXdefi';
 export { default as ConnectZerion } from './connectZerion';