+++ packages/web3-react/src/hooks/useConnectorFrontier.ts
@@ -1,0 +1,48 @@
+import invariant from 'tiny-invariant';
+import warning from 'tiny-warning';
+import { useCallback } from 'react';
+import { openWindow } from '@lido-sdk/helpers';
+import { InjectedConnector } from '@web3-react/injected-connector';
+import { useConnectors } from './useConnectors';
+import { useWeb3 } from './useWeb3';
+import { hasInjected, isFrontierProvider } from '../helpers';
+import { useForceDisconnect } from './useDisconnect';
+import { ConnectorHookArgs } from './types';
+
+type ConnectorHookResult = {
+  connect: () => Promise<void>;
+  connector: InjectedConnector;
+};
+
+const WALLET_URL = 'https://www.frontier.xyz/browser-extension/';
+
+export const useConnectorFrontier = (
+  args?: ConnectorHookArgs,
+): ConnectorHookResult => {
+  const { injected } = useConnectors();
+  const { activate } = useWeb3();
+  const { disconnect } = useForceDisconnect();
+  const onConnect = args?.onConnect;
+
+  const suggestApp = useCallback(() => {
+    try {
+      openWindow(WALLET_URL);
+    } catch (error) {
+      warning(false, 'Failed to open the link');
+    }
+  }, []);
+
+  const connect = useCallback(async () => {
+    invariant(injected, 'Connector is required');
+
+    if (hasInjected() && isFrontierProvider()) {
+      await disconnect();
+      await activate(injected);
+      onConnect?.();
+    } else {
+      suggestApp();
+    }
+  }, [injected, disconnect, activate, onConnect, suggestApp]);
+
+  return { connect, connector: injected };
+};