+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,14 @@
 # reef-knot
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@0.5.3
+  - @reef-knot/wallets-icons@0.3.0
+  - @reef-knot/web3-react@0.4.0
+
 ## 0.5.2
 
 ### Patch Changes